requirements.txt
accelerate==0.21.0
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
captum==0.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
datasets==2.13.1
decorator==5.1.1
diffusers==0.19.3
dill==0.3.6
docker-pycreds==0.4.0
executing==1.2.0
faiss-cpu==1.7.4
filelock==3.12.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
gdown==4.7.1
git-lfs==1.6
gitdb==4.0.10
GitPython==3.1.31
h5py==3.8.0
huggingface-hub==0.14.1
idna==3.4
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
invisible-watermark==0.2.0
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.13.1
kiwisolver==1.4.4
kornia==0.6.12
lit==16.0.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
ninja==1.11.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.39
protobuf==4.22.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
Pygments==2.15.1
pyparsing==3.0.9
PySocks==1.7.1
pyspng==0.1.1
python-dateutil==2.8.2
pytorch-fid==0.3.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.6.3
requests==2.29.0
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.21.1
setproctitle==1.3.2
six==1.16.0
sklearn==0.0.post4
smmap==5.0.0
soupsieve==2.5
stack-data==0.6.2
sympy==1.11.1
threadpoolctl==3.1.0
timm==0.6.13
tokenizers==0.13.3
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
unzip==1.0.0
urllib3==1.26.15
wandb==0.15.0
wcwidth==0.2.6
xxhash==3.2.0
yarl==1.9.2
zipp==3.16.0
